+++ simple_graphql_query_builder/lib/src/annotation/query_result.dart
@@ -11,11 +11,16 @@ const queryResult = QueryResult();
 class QueryResultField {
   // Custom name for query result field
   final String? name;
+
   // Should this field be omitted from the query fields
   final bool ignore;
+
+  // If the type is user defined type
+  final bool isCustomType;
 
   const QueryResultField({
     this.name,
     this.ignore = false,
+    this.isCustomType = false,
   });
 }