+++ simple_graphql_query_builder_gen/example/example.dart
@@ -3,11 +3,11 @@ import 'package:simple_graphql_query_builder/simple_graphql_query_builder.dart';
 part 'example.g.dart';
 
 @queryResult
-class Example with _$Example {
+class Example {
   final String? name;
-  @QueryResultField(ignore: false, name: 'Test')
+  @QueryResultField(ignore: false, name: 'Test2', isCustomType: true)
   final InnerExample? innerExample;
-  @QueryResultField(ignore: true)
+  @QueryResultField(isCustomType: true)
   final InnerExample innerExample2;
 
   const Example({
@@ -18,12 +18,8 @@ class Example with _$Example {
 }
 
 @queryResult
-class InnerExample with _$InnerExample {
+class InnerExample {
   final bool? innerField1;
 
   InnerExample({this.innerField1});
-}
-
-void abc() {
-  Example.ExampletoQueryResult();
 }