+++ simple_graphql_query_builder/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.2
+* Fix typo `isCutomType` to `isCustomType`
+
+## 0.2.1
+
+* Fix typo `isCutomType` to `isCustomType`
+
+## 0.2.0
+
+* Update annotation, add `isCutomType` to `QueryResultField` to generate static function instead.
+
 ## 0.1.0
 
 * Add `name` to QueryResultField