+++ simple_graphql_query_builder_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Generate functions instead, so you dont have to create a new object
+
 ## 0.1.1
 
 * Fix dependency `simple_graphql_query_builder` to `^0.1.0`